// src/global/enums/accommodationCategory.ts
export enum AccommodationCategory {
  standardSingle = "standardSingle",
  standardDouble = "standardDouble",
  standardTwin = "standardTwin",
  deluxDouble = "deluxDouble",
  studioRoom = "studioRoom",
  juniorSuit = "juniorSuit",
  executiveSuit = "executiveSuit",
  presidentailSuit = "presidentailSuit",
  royalSuit = "royalSuit",
  domitoryBunks = "domitoryBunks",
  domitoryPods = "domitoryPods",
  domitoryNoBunks = "domitoryNoBunks",
  privateSingle = "privateSingle",
  privateTwin = "privateTwin",
  privateDouble = "privateDouble",
  privateFamily = "privateFamily",
}