// src/global/enums/accommodationType.ts
export enum AccommodationType {
  apartment = "apartment",
  hostelRoom = "hostelRoom",
  hotelRoom = "hotelRoom",
  lodgeRoom = "lodgeRoom",
  motelRoom = "motelRoom",
  oficeSpace = "officeSpace",
  rental = "rental",
  shopRoom = "shopRoom",
  villa = "villa",
  manssion = "manssion",
  marketStall = "marketStall",
  supermarketSpace = "supermarketSpace",
  restaurantSpace = "restaurantSpace",
  eventsSpace = "eventsSpace",
  meetingSpace = "meetingSpace",
  conferenceSpace = "conferenceSpace",
  others = "others",
}